backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
python-decouple==3.8
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
cryptography==41.0.7
Pillow==10.3.0
groq==0.4.1
python-dotenv==1.0.0
requests==2.31.0
pdfplumber==0.11.4
setuptools>=74.1.2
drf-nested-routers==0.93.4

# Optional but recommended for full AI/OCR feature set
pytesseract==0.3.13
opencv-python-headless==4.10.0.84
numpy==2.1.2